feat: Feedback - Remove test project - Break tests into new spec file - Move getDefaultSpecFileName to util module
