--- system-tests/projects/no-specs-custom-pattern-vue/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "vue-3-app",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
-  },
-  "dependencies": {
-    "core-js": "^3.8.3",
-    "playwright-webkit": "^1.24.2",
-    "vue": "^3.2.13"
-  },
-  "devDependencies": {
-    "@cypress/vue": "4.1.0",
-    "@babel/eslint-parser": "^7.12.16",
-    "@babel/core": "^7.12.16",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^8.0.3"
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead",
-    "not ie 11"
-  ]
-}
