--- system-tests/projects/no-specs-custom-pattern-vue/src/main.js
@@ -1,4 +1,0 @@
-import { createApp } from 'vue'
-import App from './App.vue'
-
-createApp(App).mount('#app')
