--- system-tests/projects/no-specs-custom-pattern-vue/vue.config.js
@@ -1,4 +1,0 @@
-const { defineConfig } = require('@vue/cli-service')
-module.exports = defineConfig({
-  transpileDependencies: true
-})
