+++ system-tests/projects/no-specs-vue-2/cypress-custom-spec-pattern.config.js
@@ -2,8 +2,8 @@ module.exports = {
   component: {
     specPattern: 'src/specs-folder/*.cy.{js,jsx}',
     devServer: {
-      framework: "vue-cli",
-      bundler: "webpack",
-    },
-  },
-};
+      framework: 'vue-cli',
+      bundler: 'webpack'
+    }
+  }
+}
